Refactored data exporting frontend to match current code layout
